requirements.txt
Flask
torch
numpy
ultralytics
Pillow
